reorganize service package
