--- ws/src/main/scala/de/thm/ii/fbs/services/ConferenceSystemLabels.scala
@@ -1,12 +1,0 @@
-package de.thm.ii.fbs
-/**
-  * ConferenceSystemLabels holds Labels to Conference Systems
-  * @author Simon Schniedenharn
-  */
-
-object ConferenceSystemLabels {
-    /** holds label jitsi*/
-    val jitsi: String = "jitsi";
-    /** holds label bbb*/
-    val bigbluebutton: String = "bigbluebutton"
-}
