+++ ws/src/main/scala/de/thm/ii/fbs/services/labels/ConferenceSystemLabels.scala
@@ -1,0 +1,14 @@
+package de.thm.ii.fbs.services.labels
+
+/**
+  * ConferenceSystemLabels holds Labels to Conference Systems
+  *
+  * @author Simon Schniedenharn
+  */
+
+object ConferenceSystemLabels {
+  /** holds label jitsi */
+  val JITSI: String = "jitsi";
+  /** holds label bbb */
+  val BIGBLUEBUTTON: String = "bigbluebutton"
+}
