+++ ws/src/main/scala/de/thm/ii/fbs/services/labels/CourseDBLabels.scala
@@ -1,28 +1,29 @@
-package de.thm.ii.fbs.services
+package de.thm.ii.fbs.services.labels
 
 /**
   * CourseDBLabels holds all Course-Table lables
+  *
   * @author Benjamin Manns
   */
 object CourseDBLabels {
-  /** holds label courseid*/
+  /** holds label courseid */
   val courseid: String = "course_id"
-  /** holds label name*/
+  /** holds label name */
   val name: String = "course_name"
-  /** holds label description*/
+  /** holds label description */
   val description: String = "course_description"
-  /** holds label creator*/
+  /** holds label creator */
   val creator: String = "creator"
-  /** holds label standard_task_typ*/
+  /** holds label standard_task_typ */
   val standard_task_typ: String = "standard_task_typ"
-  /** holds label course_modul_id*/
+  /** holds label course_modul_id */
   val course_modul_id: String = "course_modul_id"
-  /** holds label course_semester*/
+  /** holds label course_semester */
   val course_semester: String = "course_semester"
-  /** holds label course_end_date*/
+  /** holds label course_end_date */
   val course_end_date: String = "course_end_date"
-  /** holds label personalised_submission*/
+  /** holds label personalised_submission */
   val personalised_submission: String = "personalised_submission"
-  /** holds label plagiarism_script*/
+  /** holds label plagiarism_script */
   val plagiarism_script: String = "plagiarism_script"
 }
