+++ ws/src/main/scala/de/thm/ii/fbs/services/labels/CourseParameterDBLabels.scala
@@ -1,13 +1,13 @@
-package de.thm.ii.fbs
+package de.thm.ii.fbs.services.labels
 
 /**
   * Provides DB Labels of testsystem_testfile table
   */
 object CourseParameterDBLabels {
-  /** provide course_id label **/
+  /** provide course_id label * */
   val course_id = "course_id"
-  /** provide c_param_desc **/
+  /** provide c_param_desc * */
   val c_param_desc = "c_param_desc"
-  /** provide c_param_key **/
+  /** provide c_param_key * */
   val c_param_key = "c_param_key"
 }
