+++ ws/src/main/scala/de/thm/ii/fbs/services/labels/CourseParameterUserDBLabels.scala
@@ -1,15 +1,15 @@
-package de.thm.ii.fbs
+package de.thm.ii.fbs.services.labels
 
 /**
   * Provides DB Labels of testsystem_testfile table
   */
 object CourseParameterUserDBLabels {
-  /** provide course_id label **/
+  /** provide course_id label * */
   val course_id = "course_id"
-  /** provide user_id **/
+  /** provide user_id * */
   val user_id = "user_id"
-  /** provide c_param_key **/
+  /** provide c_param_key * */
   val c_param_key = "c_param_key"
-  /** provide value **/
+  /** provide value * */
   val value = "value"
 }
