+++ ws/src/main/scala/de/thm/ii/fbs/services/labels/TestsystemLabels.scala
@@ -1,25 +1,25 @@
-package de.thm.ii.fbs.services
+package de.thm.ii.fbs.services.labels
 
 /**
   * Provides DB Labels of testsystem table
   */
 object TestsystemLabels {
-  /** provide id label **/
+  /** provide id label * */
   val id = "testsystem_id"
-  /** provide name label **/
+  /** provide name label * */
   val name = "name"
-  /** provide description label **/
+  /** provide description label * */
   val description = "description"
-  /** provide supported_formats label**/
+  /** provide supported_formats label* */
   val supported_formats = "supported_formats"
-  /** provide machine_port label **/
+  /** provide machine_port label * */
   val machine_port = "machine_port"
-  /** provide machine_ip label **/
+  /** provide machine_ip label * */
   val machine_ip = "machine_ip"
-  /** provide testfiles label **/
+  /** provide testfiles label * */
   val testfiles = "testfiles"
-  /** provide accepted_input label **/
+  /** provide accepted_input label * */
   val accepted_input = "accepted_input"
-  /** provide settings label **/
+  /** provide settings label * */
   val settings = "settings"
 }
