+++ ws/src/main/scala/de/thm/ii/fbs/services/labels/TestsystemTestfileLabels.scala
@@ -1,13 +1,13 @@
-package de.thm.ii.fbs
+package de.thm.ii.fbs.services.labels
 
 /**
   * Provides DB Labels of testsystem_testfile table
   */
 object TestsystemTestfileLabels {
-  /** provide id label **/
+  /** provide id label * */
   val id = "testsystem_id"
-  /** provide name filename **/
+  /** provide name filename * */
   val filename = "filename"
-  /** provide required label **/
+  /** provide required label * */
   val required = "required"
 }
